acr_qa_app/requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3
python-dotenv==1.0.1
pydicom==2.4.4
numpy>=2.0
opencv-python==4.10.0.84
email_validator==2.2.0
